refactor: clean, split sub views, extensions
